Hands

A Hand represents the set of tiles that a particular player has (varies in number, depending on the variety of Mahjong being played).

MahjongTiles.Hand — Type

Represents the tiles in a single player's hand.

Hands support the following operations that behave similarly to how they do in Julia Base:

iteration

push!

Scoring Functions

Hands can be checked for whether they have certain configurations/arrangements of tiles (e.g. contain no winds, contain only tiles that are completely green, contain all threes-of-a-kind). There are roughly 2 kinds:

those that describe a property of the entire hand (e.g. all sets must include the number 5), and

those that describe a smaller arrangement of tiles that must be present somewhere in the hand (e.g. two threes-in-a-row of different suits, the other tiles are not considered).

The former are generally prefixed with is_ and the latter with has_.

In a way, these are not "scoring" functions, as they don't assign a score for each configuration, but only whether the configuration applies.

MahjongTiles.is_pure — Function

is_pure(hand)

Determine whether the hand is pure (清). A hand is considered pure if there are no winds or dragons.

Example

julia> using MahjongTiles: wind, character, dragon, bamboo, circle, Hand, is_pure

julia> is_pure(Hand([wind(:east), character(3)]))
false

julia> is_pure(Hand([character(3), dragon(:green)]))
false

julia> is_pure(Hand([bamboo(8), character(3), circle(1)]))
true